refactor(architecture): remove bounded runtime cycles
