+++ ci/source-architecture-budget.json
@@ -53,10 +53,6 @@
   },
   "allowedCycles": [
     [
-      "nemoclaw/src/commands/slash.ts",
-      "nemoclaw/src/index.ts"
-    ],
-    [
       "src/lib/actions/global.ts",
       "src/lib/actions/maintenance.ts",
       "src/lib/actions/onboard.ts",
@@ -136,25 +132,12 @@
       "src/lib/onboard/policy-context-seed.ts",
       "src/lib/onboard/policy-resume-selection.ts",
       "src/lib/onboard/policy-selection.ts",
-      "src/lib/onboard/resume-provider-shim.ts",
       "src/lib/onboard/setup-inference.ts",
       "src/lib/openshell-sandbox-list.ts",
       "src/lib/sandbox/config.ts",
       "src/lib/shields/index.ts",
       "src/lib/tunnel/allowed-origins.ts",
       "src/lib/tunnel/services.ts"
-    ],
-    [
-      "src/lib/messaging/channels/teams/hooks/host-forward-port-conflict.ts",
-      "src/lib/messaging/channels/teams/hooks/index.ts",
-      "src/lib/messaging/hooks/builtins.ts",
-      "src/lib/messaging/hooks/index.ts",
-      "src/lib/messaging/persistence.ts",
-      "src/lib/messaging/plan-validation.ts"
-    ],
-    [
-      "src/lib/policy/context.ts",
-      "src/lib/policy/failure-classifier.ts"
     ]
   ],
   "maxRootFiles": {
