+++ src/lib/messaging/index.ts
@@ -7,6 +7,7 @@ export * from "./compiler";
 export * from "./diagnostics";
 export * from "./hooks";
 export * from "./host-forward";
+export * from "./hydration";
 export * from "./manifest";
 export * from "./persistence";
 export * from "./utils";
